% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mgmt.R
\name{abs_path}
\alias{abs_path}
\title{Append working directory to the paths}
\usage{
abs_path(x)
}
\arguments{
\item{x}{vector of paths}
}
\value{
a character vector the same length as x
}
\description{
Append working directory to the paths
}
