% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{append_name}
\alias{append_name}
\title{Same thing as `setNames()`}
\usage{
append_name(x, name)
}
\arguments{
\item{x}{an object}

\item{name}{the name}
}
\value{
a named vector
}
\description{
Same thing as `setNames()`
}
