% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{arg2string}
\alias{arg2string}
\title{Convert Function Arguments to Strings}
\usage{
arg2string(...)
}
\arguments{
\item{...}{Arbitrary arguments representing R language objects or symbols.}
}
\value{
A character vector where each element is the string representation of the corresponding
argument passed to the function. The order of the strings in the output matches the order of
the arguments.
}
\description{
This function takes any number of R language objects and converts their names into strings.
This is particularly useful for programming where variable names or symbols need to be used
as strings without evaluating them. It leverages \code{rlang}'s tidy evaluation framework.
}
\examples{
# returns the arguments as strings even though functions bmm() and brms() are not defined
arg2string(bmm('0.4.0'), brms('2.20.4'))
}
