% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as.function.formula}
\alias{as.function.formula}
\title{Convert formula as function}
\usage{
\method{as.function}{formula}(x)
}
\arguments{
\item{x}{the formula}
}
\value{
a function
}
\description{
Convert formula as function
}
