% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_atomic_type}
\alias{assert_atomic_type}
\title{Assert atomic type}
\usage{
assert_atomic_type(x, type, null_as_is = FALSE, NA_as_is = TRUE)
}
\arguments{
\item{x}{a vector to check}

\item{type}{the object class to check. Must have `is.CLASS()` to work.}

\item{null_as_is}{if TRUE, return NULL if x is NULL}

\item{NA_as_is}{if TRUE, return NA, if x is NA}
}
\value{
x
}
\description{
Throws error if supplied vector is not atomic or is not of expected object type
}
