% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_dir}
\alias{assert_dir}
\title{Check and expand directory path}
\usage{
assert_dir(x)
}
\arguments{
\item{x}{the directory path}
}
\value{
a character string
}
\description{
Throws an error if directory does not exist and run `file.path(x)`
}
