% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{assert_package}
\alias{assert_package}
\title{RequireNamespace() with better error message}
\usage{
assert_package(x)
}
\arguments{
\item{x}{the package}
}
\description{
RequireNamespace() with better error message
}
