% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix.R
\name{attr}
\alias{attr}
\alias{\%a\%}
\alias{\%a\%<-}
\alias{\%A\%}
\alias{\%A\%<-}
\title{Get attribute of an object}
\usage{
x\%a\%which

x\%a\%which <- value

x\%A\%which

x\%A\%which <- value
}
\arguments{
\item{x}{an object whose attributes are to be accessed.}

\item{which}{a non-empty character string specifying which attribute
    is to be accessed.}

\item{value}{an object, the new value of the attribute, or \code{NULL}
    to remove the attribute.}
}
\value{
For the extractor, the value of the attribute matched, or \code{NULL}
  if no exact match is found and no or more than one partial match is found.
}
\description{
An infix operator to get the attribute of an object (equivalent to attr(x,
"name", exact = TRUE))
}
\details{
These functions provide access to a single attribute of an object.
The replacement form causes the named attribute to take the value specified
(or create a new attribute with the value given).
}
\examples{
x <- 1
attr(x, "name") <- "John"
x\%a\%name
x\%A\%name

x\%a\%name <- "Alice"
attr(x, "name")

x <- 1:10
x\%A\%dim <- c(2,5)
x
}
