% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{available_packages}
\alias{available_packages}
\title{A character vector of available packages in the library}
\usage{
available_packages(path = .libPaths())
}
\arguments{
\item{path}{A character vector of paths to search for the package. Default is
the default library paths.}
}
\value{
A character vector of available package names in the library. If any
package versions were installed via pkg_vload(), the version will be shown as
"pkg-version"
}
\description{
A character vector of available packages in the library
}
