% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workflows.R
\name{better_rprofile}
\alias{better_rprofile}
\title{Better Rprofile}
\usage{
better_rprofile()
}
\description{
Replace the project-specific .Rprofile generated by renv with a template
that solves my headaches with VSCode and renv.
}
