% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{biplot}
\alias{biplot}
\title{Get Biplot or Triplot}
\usage{
biplot(
  x,
  choices = c(1, 2),
  scaling = 2,
  display = c("sites", "species", "biplot", "centroids"),
  ellipse = NA,
  group = NULL,
  group2 = NULL,
  sites_alpha = 1,
  sites_size = 2,
  group_as_aes = TRUE,
  species_color = "violetred",
  species_alpha = 1
)
}
\arguments{
\item{x}{an object that is supported by \code{vegan::scores()}}

\item{choices}{a vector of length 2 defining the axes to plot}

\item{scaling}{scaling argument passed to \code{vegan::scores()}}

\item{display}{a vector of characters defining what to plot}

\item{ellipse}{a character string passed to the \code{type} argument of \code{ggplot2::stat_ellipse()}. If \code{NA}, no ellipse is plotted.}

\item{group}{a vector with the same length and order as the sites data used to fit the model that is used to color the site points.}

\item{group2}{a vector with the same length and order as the sites data used to fit the model that is used to determine the shape of the site points}

\item{sites_alpha}{the transparency of the sites plotted as points}

\item{sites_size}{the size pf the site points}

\item{group_as_aes}{if \code{TRUE} (default is TRUE), \code{group} is passed as an aesthetics via \code{ggplot2::aes()}.}

\item{species_color}{the color of the species labels}

\item{species_alpha}{the transparency of the species plotted as text}
}
\value{
a ggplot object
}
\description{
Generate biplot or triplot using ggplot2.
}
