% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{collapse}
\alias{collapse}
\title{Concatenate and collapse to a single string}
\usage{
collapse(..., sep = "")
}
\arguments{
\item{...}{one or more R objects, to be converted to character vectors and
concatenated.}

\item{sep}{a character string to separate the terms.}
}
\value{
A single string.
}
\description{
Wrapper around \code{paste} with \code{collapse} argument set to \code{""}. This results in
a single string.
}
\examples{
x <- 1:10
res <- collapse(x)
identical(res, '12345678910')
}
