% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorize.R
\name{colorize}
\alias{colorize}
\title{colorize function}
\usage{
colorize(x, color = "red")
}
\arguments{
\item{x}{Text string to colorize.}

\item{color}{Color string, defaults to "red".}
}
\description{
This function allows you to color text in a Rmarkdown document. It works with
both PDF and HTML output. If used outside of a Rmarkdown document, it just
outputs the same string.
}
\examples{
colorize("text_red", color = "red")
colorize("text_white_hex", color = "#FFFFFF")
}
\keyword{color}
\keyword{text}
