% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{component_size}
\alias{component_size}
\title{Report Space Allocated for the Components of a List}
\usage{
component_size(x, units = "auto")
}
\arguments{
\item{x}{A list object}

\item{units}{A string specifying the units to use for the size. Default is "auto".}
}
\value{
A list with the same names as the input list, where each element is the size
}
\description{
Report Space Allocated for the Components of a List
}
\examples{
x <- list(a = 1:10, b = 1:100)
component_size(x)
}
