% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utils.R
\name{dev.isize}
\alias{dev.isize}
\title{Find the size of device using windows x11}
\usage{
dev.isize(expr, units = c("in", "cm", "px"))
}
\arguments{
\item{expr}{an expression to be evaluated}

\item{units}{the units of the window dimensions}
}
\description{
Uses `x11()` to interactively get (with dragging) the desired window size. Only usable for windows devices.
}
