% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{distr_fun}
\alias{distr_fun}
\title{Check whether density function, random number generation, cumulative distribution function, and quantile function are available for the specific distribution.}
\usage{
distr_fun(x)
}
\arguments{
\item{x}{a 'distr' object or a character string of the name of the distribution.}
}
\value{
a list of atomic boolean
}
\description{
Compute the mean of distribution by finding the sample mean of 10,000 draws from the distribution
}
