% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distr.R
\name{distr_mu}
\alias{distr_mu}
\title{Compute the mean of distribution via simulation}
\usage{
distr_mu(x)
}
\arguments{
\item{x}{a 'distr' object}
}
\value{
an atomic numeric vector
}
\description{
Compute the mean of distribution by finding the sample mean of 10,000 draws from the distribution
}
