% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{drop_empty_list}
\alias{drop_empty_list}
\title{Find each empty list element then drop them}
\usage{
drop_empty_list(x)
}
\arguments{
\item{x}{list of a bunch of nested lists}
}
\value{
a list
}
\description{
Find each empty list element then drop them
}
