% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{extract_pkg_fun_calls}
\alias{extract_pkg_fun_calls}
\title{Extract all calls to a function from package code}
\usage{
extract_pkg_fun_calls(pkg, fun)
}
\arguments{
\item{pkg}{Name of the package from which to extract function calls}

\item{fun}{Name of the function to extract calls to}
}
\value{
A character vector with the extracted function calls. Each value is
of the form "fun(arg1=value1, arg2=value2, ...)"
}
\description{
Given a function name and a package, this function will extract all calls to
the said function from the package. The function can be any function name,
either imported or defined in the package.
}
\examples{
extract_pkg_fun_calls("utils", "getOption")
}
