% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{fake_pkg}
\alias{fake_pkg}
\title{Find fake package generated by `vmisc::load_all2()` or `pkgload::load_all()`}
\usage{
fake_pkg(fake_pkg_root_path = getwd())
}
\arguments{
\item{fake_pkg_root_path}{the root path of where the fake package is installed at}
}
\value{
the name of the fake package
}
\description{
Find fake package generated by `vmisc::load_all2()` or `pkgload::load_all()`
}
