% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{fancy_linear}
\alias{fancy_linear}
\alias{fancy_scientific}
\alias{fancy_scientificb}
\alias{scale_x_ta}
\alias{scale_xy_log}
\title{Fancy ggplot scales}
\usage{
fancy_linear(l)

fancy_scientific(l)

fancy_scientificb(l)

scale_x_ta(name = ggplot2::waiver(), ...)

scale_xy_log(name = ggplot2::waiver(), axis = c("x", "y"))
}
\arguments{
\item{l}{character string}

\item{name}{The name of the scale. Used as the axis or legend title. If `waiver()`, the default, the name of the scale is taken from the first mapping used for that aesthetic. If NULL, the legend title will be omitted.}

\item{...}{additional arguments}

\item{axis}{"x" log the x-axis; "y" log the y-axis}
}
\value{
an expression
}
\description{
variously stolen from other places
}
