% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{flatten_mat_name}
\alias{flatten_mat_name}
\title{Turn named matrix into a named vector}
\usage{
flatten_mat_name(x)
}
\arguments{
\item{x}{matrix}
}
\value{
a named vector
}
\description{
Turn named matrix into a named vector
}
