% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{FOSM2}
\alias{FOSM2}
\title{Multivariate delta method}
\usage{
FOSM2(x, vcv, trans)
}
\arguments{
\item{x}{the expectation of x}

\item{vcv}{the variance covariance matrix}

\item{trans}{a list of functions of the transformation}
}
\value{
a variance covariance matrix
}
\description{
Approximate the variance covariance matrix of transformed random variables via first order Taylor approximation
}
