% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utils.R
\name{geom_lineeq}
\alias{geom_lineeq}
\title{Add fitted line and equation geom layer}
\usage{
geom_lineeq(
  x,
  method = "MA",
  rr.digits = 3,
  coef.digits = 3,
  labels = c("eq", "R2"),
  ...
)
}
\arguments{
\item{x}{a ggplot object}

\item{method}{the method of fitting ("MA": major axis, "OLS": ordinary least squares, "SMA": standardized major axis, "RMA": range major axis)}

\item{rr.digits, coef.digits}{the number of digits to display}

\item{labels}{a vector of what to label. ("eq", "R2", "P", "n")}

\item{...}{additional arguments}
}
\value{
a ggplot object
}
\description{
a wrapper for `ggpmisc::stat_ma_line()` and `ggpmisc::stat_ma_eq()`
}
