% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_time.R
\name{get_time}
\alias{get_time}
\title{get_time function}
\usage{
get_time(datetime = lubridate::now())
}
\arguments{
\item{datetime}{Date-time column/variable. Defaults to lubridate::now().}
}
\description{
This function allows you to extract the time elements from a date-time.
}
\examples{
get_time()
}
\keyword{date}
\keyword{time}
