% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{gggradfield}
\alias{gggradfield}
\title{Plot gradient field}
\usage{
gggradfield(
  data,
  aggregate = 2,
  colours = c("#B2182B", "#E68469", "#D9E9F1", "#ACD2E5", "#539DC8", "#3C8ABE",
    "#2E78B5")
)
}
\arguments{
\item{data}{a data with x, y, z}

\item{aggregate}{bin size to aggregate data}

\item{colours}{colors of gradient field}
}
\value{
a ggplot
}
\description{
Plot gradient field
}
