% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr.R
\name{\%a\%}
\alias{\%a\%}
\title{Get attribute of an object}
\usage{
object \%a\% a
}
\arguments{
\item{object}{an object whose attributes are to be accessed.}

\item{a}{an expression specifying the attribute to be accessed.}
}
\value{
For the extractor, the value of the attribute matched, or NULL if no exact match is found and no or more than one partial match is found.
}
\description{
Get attribute of an object
}
\examples{
x <- 1
attr(x, "name") <- "ven"
x\%a\%name
}
