% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attr.R
\name{\%+\%}
\alias{\%+\%}
\title{Shortcut for paste0}
\usage{
a \%+\% b
}
\arguments{
\item{a}{an R object to be converted to a character vector.}

\item{b}{an R object to be converted to a character vector.}
}
\value{
String concatenation of a and b
}
\description{
Shortcut for paste0
}
\examples{
"name" \%+\% 1:10
}
