% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{gsub_element_wise}
\alias{gsub_element_wise}
\title{Perform `gsub()` by element}
\usage{
gsub_element_wise(
  pattern,
  replacement,
  x,
  ignore.case = FALSE,
  perl = FALSE,
  fixed = FALSE,
  useBytes = FALSE
)
}
\arguments{
\item{pattern}{see `?gsub()`}

\item{replacement}{see `?gsub()`}

\item{x}{see `?gsub()`}

\item{ignore.case}{see `?gsub()`}

\item{perl}{see `?gsub()`}

\item{fixed}{see `?gsub()`}

\item{useBytes}{see `?gsub()`}
}
\value{
a character string
}
\description{
Perform `gsub()` by element
}
