% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{has_function}
\alias{has_function}
\title{Check whether function exists and is function}
\usage{
has_function(x)
}
\arguments{
\item{x}{the object}
}
\value{
TRUE or FALSE
}
\description{
Check whether function exists and is function.
}
