% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mgmt.R
\name{init_dir}
\alias{init_dir}
\title{Create directory if it doesn't exists}
\usage{
init_dir(root_path, dir_name)
}
\arguments{
\item{root_path}{path to where the directory should be located}

\item{dir_name}{name of the directory}
}
\description{
Create directory if it doesn't exists
}
