% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.between}
\alias{is.between}
\title{Check if value is between range}
\usage{
is.between(x, range, inclusive = FALSE)
}
\arguments{
\item{x}{a vector of numeric values}

\item{range}{a numeric vector of length 2 indicating the range}

\item{inclusive}{should the bound be included (default is FALSE)?}
}
\value{
a logical vector the same length as x
}
\description{
Check if value is between range
}
