% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.multi_response}
\alias{is.multi_response}
\title{Check if brmsfit is multivariate}
\usage{
is.multi_response(model)
}
\arguments{
\item{model}{a model of class 'brmsfit'}
}
\value{
a boolean
}
\description{
Check if brmsfit is multivariate
}
