% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.null_na}
\alias{is.null_na}
\title{Check if is null, else, return a vector of is.na(x)}
\usage{
is.null_na(x)
}
\arguments{
\item{x}{the thing to check}
}
\value{
a logical vector
}
\description{
a safer version of is.na(x) that also deals with NULL
}
