% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.numbers_only}
\alias{is.numbers_only}
\title{Returns TRUE if the string has only numbers}
\usage{
is.numbers_only(x)
}
\arguments{
\item{x}{a vector of character string}
}
\value{
a logical vector the same length as x
}
\description{
Returns TRUE if the string has only numbers
}
