% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{is.odd}
\alias{is.odd}
\title{Check if value is odd}
\usage{
is.odd(x)
}
\arguments{
\item{x}{a vector of numeric values}
}
\value{
a logical vector the same length as x
}
\description{
Check if value is odd
}
