% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mgmt.R
\name{is_relative_path}
\alias{is_relative_path}
\title{Check if path is relative to current working directory}
\usage{
is_relative_path(x)
}
\arguments{
\item{x}{vector of paths}
}
\value{
a logical vector the same length as x
}
\description{
Check if path is relative to current working directory
}
