% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{isTRUE_elementwise}
\alias{isTRUE_elementwise}
\alias{isFALSE_elementwise}
\title{Vectorized isTRUE() or isFALSE()}
\usage{
isTRUE_elementwise(x)

isFALSE_elementwise(x)
}
\arguments{
\item{x}{a vector}
}
\value{
a logical vector
}
\description{
Vectorized isTRUE() or isFALSE()
}
