% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{keep_rowname}
\alias{keep_rowname}
\title{Set rowname as the first column}
\usage{
keep_rowname(x, rn = "rn")
}
\arguments{
\item{x}{matrix or data.frame}

\item{rn}{name of the rowname column}
}
\value{
a matrix or data.frame
}
\description{
Set rowname as the first column
}
