% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{lapply_name}
\alias{lapply_name}
\title{lapply with append_name()}
\usage{
lapply_name(x, FUN, ...)
}
\arguments{
\item{x}{a vector}

\item{FUN}{a function to be applied to each element of the vector x}

\item{...}{additional arguments}
}
\value{
a list
}
\description{
lapply() but with names(z) <- x
}
