% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_format.R
\name{left_align_text}
\alias{left_align_text}
\title{Left align text}
\usage{
left_align_text(x)
}
\arguments{
\item{x}{a vector of character strings}
}
\value{
a vector of character strings
}
\description{
Left align text
}
