% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{map_if_depth}
\alias{map_if_depth}
\title{Conditional map to each nest element of a nested list}
\usage{
map_if_depth(x, .f, .p, .f2 = identity)
}
\arguments{
\item{x}{a vector}

\item{.f}{a function to be evaluated on each element of x if .p is TRUE}

\item{.p}{a condition to test on each element of x}

\item{.f2}{a function to be evaluated on each element of x if .p is FALSE. Default to identity}
}
\value{
the output of the expression
}
\description{
Apply a function .f to each nested element of a list x that satisfy the condition .p(.x).
}
