% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vis_utils.R
\name{marginal_effects}
\alias{marginal_effects}
\title{Generate marginal effect data frame for plotting}
\usage{
marginal_effects(model, terms, n = 300, ci = 0.95, ...)
}
\arguments{
\item{model}{the model object}

\item{terms}{the terms for which to compute the marginal effects. Has the same behavior as in `sjPlot::plot_model()`}

\item{n}{specifies the density of points along `terms`}

\item{ci}{the interval of upper and lower bounds}

\item{...}{additional arguments passed to `predict()`}
}
\value{
a data.frame
}
\description{
Generate marginal effect data frame for plotting
}
