% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_table.R
\name{model_table}
\alias{model_table}
\title{model_table function}
\usage{
model_table(run, dir, col = NULL)
}
\arguments{
\item{run}{The run number}

\item{dir}{Path to the run directory}

\item{col}{Columns to print}
}
\value{
A data frame.
}
\description{
Gather data from a NONMEM run using xpose::xpose_data(), and output it into a
table. By default, it outputs the run number, the reference run number, the
contents of \verb{$PROBLEM} (model description), the number of parameters, the OFV,
the dOFV from the reference run, and the termination message.
Credit: Estelle Chasseloup
}
\examples{
\dontrun{
model_table(
  run = 163,
  dir = "/path/to/run/directory"
)
}
}
\keyword{nonmem}
\keyword{results}
\keyword{table}
\keyword{xpose}
