% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{moment}
\alias{moment}
\title{Compute raw moment}
\usage{
moment(x, p)
}
\arguments{
\item{x}{a vector of samples}

\item{p}{an atomic numeric vector of the pth moment}
}
\value{
an atomic vector
}
\description{
compute the pth raw moment
}
