% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_mgmt.R
\name{move_files}
\alias{move_files}
\title{Move files with `file.rename()` with a progress bar}
\usage{
move_files(from_dir, to_dir, file_names, pb = FALSE)
}
\arguments{
\item{from_dir}{path where the files should be looked for}

\item{to_dir}{path to where the files should be moved}

\item{file_names}{name of files to be moved}

\item{pb}{if TRUE, show a progress bar}
}
\value{
a logical vector of whether moving each file succeeded
}
\description{
Move files with `file.rename()` with a progress bar
}
