% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_fun.R
\name{n_fun}
\alias{n_fun}
\title{n_fun function}
\usage{
n_fun(x)
}
\arguments{
\item{x}{smth}
}
\description{
Output the number of data points in a box plot.
For use within \code{ggplot2::stat_summary()}.
}
\examples{
\dontrun{
diamonds \%>\%
ggplot(aes(x = cut, y = depth)) +
 geom_boxplot() +
 stat_summary(fun.data = n_fun, geom = "text", size = 5, color = "blue")
)
}
}
\keyword{box}
\keyword{number}
\keyword{plot}
