% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_objects.R
\name{nlist}
\alias{nlist}
\title{Create Named List from Arguments}
\usage{
nlist(...)
}
\arguments{
\item{...}{Arbitrary arguments to be included in the list. These can be named or unnamed.
Unnamed arguments will be named based on their variable names.}
}
\value{
A list where each element corresponds to an argument passed to the function.
Elements of the list are named based on either their original names or the names of
the variables passed as arguments.
}
\description{
This function creates a named list from its arguments. If the arguments are named,
those names are used in the resulting list. If some arguments are unnamed, the variable
names themselves are used as names in the list. This can be useful for creating lists
where the names are important for later indexing or manipulation, and ensures all
elements in the list have names.
}
\examples{
var1 <- 1
var2 <- 1:10
# This will return a list with names: c("a", "b", "var1", "var2")
nlist(a = 1, b = 2, var1, var2)
}
