% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{null_to_NA}
\alias{null_to_NA}
\title{Convert NULL value to NA vector of length 'len'}
\usage{
null_to_NA(x, len = 1, error_NA = FALSE)
}
\arguments{
\item{x}{a vector which may be NULL}

\item{len}{number of NA to return when value is NULL}

\item{error_NA}{if TRUE, if the vector contains errors, do the same thing as if the value of x is NULL.}
}
\value{
a vector
}
\description{
Convert NULL value to NA vector of length 'len'
}
