% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omit.R, R/zS3_methods.R
\name{omit_zero.list}
\alias{omit_zero.list}
\alias{omit_zero.default}
\alias{omit_zero}
\title{Drop the element if it is completely zero}
\usage{
\method{omit_zero}{list}(x)

\method{omit_zero}{default}(x)

omit_zero(x, ...)
}
\arguments{
\item{x}{a vector}

\item{...}{additional arguments}
}
\description{
Drop the element if it is completely zero
}
