% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assert.R
\name{overlaps_zero}
\alias{overlaps_zero}
\title{Check if range of vector contains zero}
\usage{
overlaps_zero(..., return_star = FALSE)
}
\arguments{
\item{...}{additional numeric vectors}

\item{return_star}{if TRUE, return '*' or "" in place of TRUE or FALSE}
}
\value{
a logical or logical vector the same length as ...
}
\description{
Check if range of vector contains zero
}
