% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{package_version2}
\alias{package_version2}
\title{Find package version}
\usage{
package_version2(x)
}
\arguments{
\item{x}{the name of the package}
}
\value{
the package version
}
\description{
same as `utils::packageVersion()` but with error handling
}
