% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{parse_pkg_version}
\alias{parse_pkg_version}
\title{Parse package name and version from a pkg('verions') call}
\usage{
parse_pkg_version(...)
}
\arguments{
\item{...}{a number of calls to objects of type pkg('version') where pkg is
the package name and version is the version number}
}
\value{
A list with two elements: names and versions. The names are the package
names and the versions are of class 'package_version'. If the version is not
specified, the version will be NA.
}
\description{
Parse package name and version from a pkg('verions') call
}
\examples{
parse_pkg_version(brms("2.20.4"), bmm("0.4-0"), utils)
}
