% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{pb_par_lapply}
\alias{pb_par_lapply}
\title{lapply with multi-session parallel evaluation using doParallel backend}
\usage{
pb_par_lapply(
  x,
  FUN,
  cores = 1,
  ...,
  inorder = TRUE,
  export_fun_only = TRUE,
  fake_pkg_root_path = getwd()
)
}
\arguments{
\item{x}{a vector}

\item{FUN}{a function to be evaluated on each element of x}

\item{cores}{number of parallel sessions used for function evaluation of vector elements.}

\item{...}{additional arguments passed to FUN}

\item{inorder}{If TRUE (default), each function evaluation of vector elements must be in the same order as they are supplied}

\item{export_fun_only}{If TRUE (default), only functions in the global environment are exported to parallel sessions}

\item{fake_pkg_root_path}{the path of the fake package}
}
\value{
a list
}
\description{
Same deal as `lapply()` but allows for parallel computing, progress bar, and fake package compatability handling.
}
