% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pkg_switch_default}
\alias{pkg_switch_default}
\title{Switch the default version of a package}
\usage{
pkg_switch_default(pkg, new_default_version, path = .libPaths())
}
\arguments{
\item{pkg}{The name of the package}

\item{new_default_version}{The version of the package to be set as the default}

\item{path}{A character vector of paths to search for the package. Default is
the default library paths.}
}
\value{
Invisible \code{TRUE} if the default version was successfully switched, \code{FALSE}
otherwise.
}
\description{
\code{\link[=pkg_switch_default]{pkg_switch_default()}} renames the folder of the default version of a package
to pkg-version and renames the folder of the new default version to pkg. This
allows the user to switch between different versions of the same package that
are loaded by the default call to \code{library()}.
}
\examples{
\dontrun{
# install two versions of the xfun package
pkg_vload(stringr, stringr('1.0.0'))
# switch the default version of xfun to 0.2.0
pkg_switch_default("stringr", "1.0.0")
}
}
