% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{pkg_vavailable}
\alias{pkg_vavailable}
\title{Check if a specific package version is available in the library}
\usage{
pkg_vavailable(..., path = .libPaths(), exact = TRUE)
}
\arguments{
\item{...}{One or more calls to the package name with version (if desired)
specified in parantheses. E.g. brms("2.14.4") or brms or "brms"}

\item{path}{A character vector of paths to search for the package. Default is
the default library paths.}

\item{exact}{Logical. If \code{TRUE}, the function will only return \code{TRUE} if the
exact version is available. If \code{FALSE}, the function will return \code{TRUE} if
the version is available or if a higher version is available. Default is \code{TRUE}.}
}
\value{
a named list with the following elements:
\itemize{
\item available: A logical vector indicating whether the package is available
\item pkg_name: The name of the package
\item pkg_version: The version of the package in the library
\item pkg_version_specified: The version of the package specified in the call to pkg('version')
\item pkg_folder: The folder name of the package in the library
}
}
\description{
\code{\link[=pkg_vavailable]{pkg_vavailable()}} is an alternative to \code{\link[xfun:pkg_attach]{xfun::pkg_available()}} that checks
for a specific version of the package rather than a minimal version. If the
version is not specified, the function will check for the default version of
the package.
}
\details{
To check for a specific version, the function assumes that this
version was installed using pkg_load(pkg(version)), which has
created a folder named "pkg-version" in the library.
}
\examples{
\dontrun{
pkg_vavailable(utils)
pkg_vavailable(xfun("0.1.0"))
pkg_vavailable(utils, brms("2.14.4"), xfun("0.1.0"))

# compare with xfun::pkg_available()
xfun::pkg_available("xfun", "0.1.0") # returns TRUE
}
}
