% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_moments}
\alias{plot_moments}
\title{Plot raw moments against the first moment}
\usage{
plot_moments(
  data_list,
  choice = c(2:4),
  plot = TRUE,
  center = FALSE,
  rr.digits = 3,
  coef.digits = 3,
  method = "MA"
)
}
\arguments{
\item{data_list}{a list of numeric sample values}

\item{choice}{a vector of pth moments}

\item{plot}{a logical value indicating whether to plot the results.}

\item{center}{whether to compute the central moment instead of the raw moments (default is FALSE)}

\item{rr.digits, coef.digits}{number of digits to display}

\item{method}{the method of fitting ("MA": major axis, "OLS": ordinary least squares, "SMA": standardized major axis, "RMA": range major axis)}
}
\value{
a data.frame which is used to generate the ggplot
}
\description{
Plot raw moments against the first moment. Useful for testing the prediction of scale collapse (i.e., when the slope of the relationship between the pth raw moment and the first raw moment is p).
}
