% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_rank_size}
\alias{plot_rank_size}
\title{Rank size plot}
\usage{
plot_rank_size(x, omit_zero = TRUE, add = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{a vector of numeric values greater than 0}

\item{omit_zero}{If TRUE (default), remove 0 from x.}

\item{add}{a logical value indicating whether to overlay points upon an existing plot}

\item{plot}{a logical value indicating whether to plot the results.}

\item{...}{additional arguments passed to \code{points()} or \code{plot()}}
}
\value{
a data.frame of the rank and log values
}
\description{
Generate a rank size plot
}
