% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_survival}
\alias{plot_survival}
\title{Survival Plot}
\usage{
plot_survival(x, add = FALSE, plot = TRUE, ...)
}
\arguments{
\item{x}{a vector of numeric values greater than 0}

\item{add}{a logical value indicating whether to overlay points upon an existing plot}

\item{plot}{a logical value indicating whether to plot the results. Set to \code{FALSE} to just getting the empirical cumulative distribution function data.frame.}

\item{...}{additional arguments passed to \code{points()} or \code{plot()}}
}
\value{
a data.frame of the empirical cumulative distribution function
}
\description{
Generate a log-log survival plot of positive only sample. Can be useful for visualizing the tail of the distribution.
}
\examples{
x <- rlnorm(10000,1,1)
plot_survival(x)

x2 <- rlnorm(10000,1,1.1)
plot_survival(x2, add = TRUE, col = "green", pch = 19)

}
\references{
Aban, I. B., M. M. Meerschaert, and A. K. Panorska. 2006. Parameter Estimation for the Truncated Pareto Distribution. Journal of the American Statistical Association 101:270–277.
}
