% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{pooled_SD}
\alias{pooled_SD}
\title{Compute pooled SD}
\usage{
pooled_SD(v, n)
}
\arguments{
\item{v}{variance of group i}

\item{n}{sample size of group i}
}
\value{
the pooled SD
}
\description{
compute pooled standard deviation
}
