% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R, R/zS3_methods.R
\name{posterior_epred.gam}
\alias{posterior_epred.gam}
\alias{posterior_epred.brmsfit}
\alias{posterior_epred.nls}
\alias{posterior_epred}
\title{Generate posterior draws from model}
\usage{
\method{posterior_epred}{gam}(
  x,
  newdata = NULL,
  ndraws = 100,
  unconditional = FALSE,
  re_formula = NA,
  scale = c("response", "link"),
  ...
)

\method{posterior_epred}{brmsfit}(
  x,
  newdata = NULL,
  re_formula = NULL,
  resp = NULL,
  dpar = NULL,
  nlpar = NULL,
  ndraws = NULL,
  draw_ids = NULL,
  sort = FALSE,
  scale = c("response", "link"),
  ...
)

\method{posterior_epred}{nls}(
  x,
  newdata = NULL,
  ndraws = 100,
  re_formula = NA,
  scale = c("response", "link"),
  ...
)

posterior_epred(
  x,
  newdata = NULL,
  ndraws,
  re_formula = NA,
  scale = c("response", "link"),
  ...
)
}
\arguments{
\item{x}{the model object}

\item{newdata}{a data.frame of newdata for the model to predict from}

\item{ndraws}{a positive integer indicating the number of draws from the model}

\item{unconditional}{if TRUE then the smoothing parameter uncertainty corrected covariance matrix is used, when available, otherwise the covariance matrix conditional on the estimated smoothing parameters is used.}

\item{re_formula}{if NA, ignore random effects.}

\item{scale}{return the predictions on the link scale ("link") or response scale ("response"). Default is "response".}

\item{...}{additional arguments}

\item{resp}{passed to `brms::prepare_predictions()`}

\item{dpar}{passed to `rstantools::posterior_epred()`}

\item{nlpar}{passed to `rstantools::posterior_epred()`}

\item{draw_ids}{passed to `brms::prepare_predictions()`}

\item{sort}{passed to `rstantools::posterior_epred()`}
}
\value{
a matrix of ndraws X nobs
}
\description{
Generate posterior draws from model.
}
