% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{prepare_newdata}
\alias{prepare_newdata}
\title{Takes a model and specified predictor terms and format newdata used for predictions.}
\usage{
prepare_newdata(model, terms = NULL, n = 300)
}
\arguments{
\item{model}{the model object}

\item{terms}{the terms for which to compute the marginal effects. Has the same behavior as in `sjPlot::plot_model()`}

\item{n}{specifies the number of draws.}
}
\description{
Covariates not specified in terms are set to: (numeric = the mean), (otherwise: set to unique entries)
}
