% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_format.R
\name{pretty_size}
\alias{pretty_size}
\title{Find object size in pretty format}
\usage{
pretty_size(x)
}
\arguments{
\item{x}{the object}
}
\value{
pretty object size
}
\description{
Find object size in pretty format
}
