% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_utils.R
\name{recompile}
\alias{recompile}
\title{Recompile C++ code}
\usage{
recompile(package = fake_pkg())
}
\arguments{
\item{package}{the name of the package to recompile C++ code}
}
\description{
Recompile C++ code for a specified package with `pkgload`
}
