% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packages.R
\name{require_pkg}
\alias{require_pkg}
\title{Check Required Packages and Their Versions}
\usage{
require_pkg(..., message_prefix = "Please install the following packages:")
}
\arguments{
\item{...}{Variable arguments representing required package names and,
optionally, their minimum versions. The versions should be specified
immediately after the package names, in the format \code{packageName(version)}.}

\item{message_prefix}{A character string to be displayed before the message
if the requirements are not met.}
}
\value{
This function does not return a value. Instead, it will stop
the execution and display a message if the requirements are not met.
}
\description{
This function checks if the required R packages are available and if their
versions meet the specified minimum requirements. It will stop the execution
and display a message if any required package is missing or does not meet
the version requirement.
}
\examples{
\dontrun{
# Check if 'dplyr' and 'ggplot2' are installed (any versions):
require_pkg(dplyr, ggplot2)

# Check if 'dplyr' (version 1.0.0 or higher) and 'ggplot2' (version 8.3.0 or higher) are installed:
require_pkg(dplyr('1.0.0'), ggplot2('8.3.0'))
}
}
