% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reverse_names}
\alias{reverse_names}
\title{Reverse name and value}
\usage{
reverse_names(x)
}
\arguments{
\item{x}{a named vector}
}
\value{
a named character vector
}
\description{
Reverse name and value
}
