% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_format.R
\name{hms_runtime}
\alias{hms_runtime}
\alias{hms_format}
\alias{hm_format}
\title{Format seconds to readable format}
\usage{
hms_runtime(x)

hms_format(x)

hm_format(x)
}
\arguments{
\item{x}{seconds}
}
\value{
a character string
}
\description{
Format seconds to readable format
}
