% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_format.R
\name{sigfig}
\alias{sigfig}
\title{Find the significant figures}
\usage{
sigfig(x, digits = 2)
}
\arguments{
\item{x}{a vector of numeric values}

\item{digits}{an atomic integer indicating the number of significant figures.}
}
\value{
a character vector
}
\description{
Find the significant figures
}
