% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_checks.R, R/zS3_methods.R
\name{simulateResiduals.brmsfit}
\alias{simulateResiduals.brmsfit}
\alias{simulateResiduals.nls}
\alias{simulateResiduals.default}
\alias{simulateResiduals}
\title{Create simulated residuals}
\usage{
\method{simulateResiduals}{brmsfit}(x, integer = NULL, plot = FALSE, asFactor = FALSE, ...)

\method{simulateResiduals}{nls}(x, asFactor = FALSE, plot = FALSE, integer = FALSE, ...)

\method{simulateResiduals}{default}(x, plot = FALSE, asFactor = NULL, ...)

simulateResiduals(x, plot = FALSE, asFactor, ...)
}
\arguments{
\item{x}{the model object}

\item{integer}{is the response an integer?}

\item{plot}{if TRUE, plot the simulated residuals}

\item{asFactor}{if TRUE, bin the residuals.}

\item{...}{additional arguments}
}
\description{
see `?DHARMa::simulateResiduals()`.
}
