% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{spaghetti}
\alias{spaghetti}
\title{Get posterior draws on the response or link scale based on specified terms.}
\usage{
spaghetti(
  model,
  terms,
  n = 100,
  along_n = 300,
  newdata2 = NULL,
  transFUN = NULL,
  scale = c("response", "link")
)
}
\arguments{
\item{model}{model object}

\item{terms}{the terms for which to compute the marginal effects. Has the same behavior as in `sjPlot::plot_model()`}

\item{n}{specifies the number of draws.}

\item{along_n}{specifies the density of points along `terms`}

\item{newdata2}{passes extra data not specified in `terms`. Can be trial size or random effects.}

\item{transFUN}{passes a function used to transform the posterior draws. Defaults to identity.}

\item{scale}{return the predictions on the link scale ("link") or response scale ("response"). Default is "response".}
}
\description{
Get posterior draws on the response or link scale based on specified terms. Specific values of terms can be specified with the `[]` syntax. e.g., `[2,3,4]` or `[c(1:3)]`. Supports all model object class with a valid `posterior_epred()` method.
}
