% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pretty_format.R
\name{stars_pval}
\alias{stars_pval}
\title{Return stars based on p value.}
\usage{
stars_pval(x)
}
\arguments{
\item{x}{a vector of p values}
}
\value{
a character vector of stars
}
\description{
Return stars based on p value like in `summary.lm()`
}
