% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stop2}
\alias{stop2}
\alias{warning2}
\title{Wrappers around \code{stop} and \code{warning} that do not print the call stack}
\usage{
stop2(...)

warning2(...)
}
\arguments{
\item{...}{zero or more objects which can be coerced to character (and which
are pasted together with no separator) or a single condition object}
}
\value{
Stops execution and prints the error message
}
\description{
Wrappers around \code{stop} and \code{warning} that do not print the call stack
}
