% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{str_extract_nested_balanced}
\alias{str_extract_nested_balanced}
\title{Extract a substring defined by a prefix, and a matched opening and closing
character}
\usage{
str_extract_nested_balanced(string, prefix, opening = "(", closing = ")")
}
\arguments{
\item{string}{A character vector}

\item{prefix}{A string to match at the beginning of the substring}

\item{opening}{A character to match as the opening character}

\item{closing}{A character to match as the closing character}
}
\value{
A character vector with the extracted substring
}
\description{
Used to, for example, extract a function call from code, ignoring paranthesis
within function arguments
}
\details{
There will be a match only if the function is not used as text in
quotes immediately before the function call. For example, in the string
"something myFunction(x = 1, y = mean(x)) otherFunction()", the function
call to \code{myFunction} will be matched, but the function call in the string
"var = \"myFunction(x = 1, y = mean(x))\"" or print("myFunction(x = 1, y =
mean(x))") will not. This ensures that we do not retrieve examples or
instructions about usage
}
\examples{

x <- "something myFunction(x = 1, y = mean(x)) otherFunction()"
str_extract_nested_balanced(x, "myFunction", "(", ")")
}
