% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_attributes}
\alias{strip_attributes}
\title{Remove all attributes of an object except those specified as protected}
\usage{
strip_attributes(x, protect = c("names", "row.names", "class"))
}
\arguments{
\item{x}{an R object}

\item{protect}{a character vector of attribute names to keep. Default is
\code{c("names", "row.names", "class")}, which are the attributes that a
data.frame has by default.}
}
\value{
An R object with all attributes removed except those specified in
\code{protect}.
}
\description{
Remove all attributes of an object except those specified as protected
}
\examples{
x <- data.frame(a = 1:10, b = 11:20)
attr(x, "remove_me") <- "I want to be removed"
attributes(x)
x <- strip_attributes(x, protect = c("names", "row.names", "class"))
attributes(x)
}
