% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{suppressExtraMessages}
\alias{suppressExtraMessages}
\title{Suppress duplicate warnings or messages}
\usage{
suppressExtraMessages(expr)
}
\arguments{
\item{expr}{an expression}
}
\value{
output of expr
}
\description{
Collapse duplicate messages.
}
