% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{text_col}
\alias{text_col}
\title{Change the text color depending on rstudio theme dark mode or light mode}
\usage{
text_col(x)
}
\arguments{
\item{x}{a character string}
}
\value{
a character string
}
\description{
stolen from `tidyverse`
}
