% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unmelt}
\alias{unmelt}
\title{Reshape an array in long format to an array}
\usage{
unmelt(x, colname = "val")
}
\arguments{
\item{x}{a matrix or data.frame with named columns. The value of each cell is stored as a 'val' column.}

\item{colname}{optional argument to change the name of the value column to look for. Default is 'val'.}
}
\value{
an array
}
\description{
Reshape an array in the long format where each row correspond to the index of a value in the nth dimension and the value is stored in the 'val' column into an nth dimensional array.
}
