% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{unscale}
\alias{unscale}
\alias{unscalelog}
\title{Unscale variable}
\usage{
unscale(x)

unscalelog(logx)
}
\arguments{
\item{x, logx}{a vector}
}
\value{
a function
}
\description{
Generate a function from a vector that can back transform a Z-score transformed vector to its original scale.
}
