% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{vcov.nls}
\alias{vcov.nls}
\title{Variance covariance matrix}
\usage{
\method{vcov}{nls}(object)
}
\arguments{
\item{object}{an object of class 'nls'}
}
\value{
the variance covariance matrix
}
\description{
get the VCV matrix
}
