% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vmisc-package.R
\docType{package}
\name{vmisc-package}
\alias{vmisc}
\alias{vmisc-package}
\title{vmisc: Vincent's misc functions}
\description{
Vincent's misc functions to be used across projects.
}
\author{
\strong{Maintainer}: Vincent Pan \email{vsbpan@gmail.com} (\href{https://orcid.org/0000-0001-9892-7805}{ORCID})

Authors:
\itemize{
  \item Vincent Pan \email{vsbpan@gmail.com} (\href{https://orcid.org/0000-0001-9892-7805}{ORCID})
}

}
\keyword{internal}
