% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do.R
\name{with_timeout}
\alias{with_timeout}
\title{Timeout wrapper}
\usage{
with_timeout(
  expr,
  substitute = TRUE,
  envir = parent.frame(),
  timeout,
  cpu = timeout,
  elapsed = timeout,
  ...
)
}
\arguments{
\item{expr}{an expression to be evaluated}

\item{substitute}{If true (default), substitute the expression}

\item{envir}{In what environment to evaluate the expression? Default to parent environment.}

\item{timeout}{How many seconds to wait before timing out}

\item{cpu, elapsed}{double (of length one). Set a limit on the total or elapsed CPU time in seconds, respectively.}

\item{...}{additional arguments not passed to anything at the moment.}
}
\value{
the output of the expression
}
\description{
Evaluate an expression within a limited time. If timed out, throws an error.
}
