% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{zbind}
\alias{zbind}
\title{Bind matrices together into an array along the z dimension}
\usage{
zbind(...)
}
\arguments{
\item{...}{matrices of the same width and height}
}
\value{
a three dimensional array
}
\description{
Bind matrices together into an array along the z dimension
}
