% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infix_functions.R
\name{\%+\%}
\alias{\%+\%}
\title{Shortcut for paste0}
\usage{
a \%+\% b
}
\arguments{
\item{a}{an R object to be converted to a character vector.}

\item{b}{an R object to be converted to a character vector.}
}
\value{
String concatenation of a and b
}
\description{
Equivalent to python's string concatenation via + operator
}
\examples{
# adding a string and a vector
"name" \%+\% 1:10

# adding vector variables and strings
names <- c("John", "Sarah")
ages <- c(34,23)
res <- names \%+\% " is " \%+\% ages \%+\% " years old"
identical(res, c("John is 34 years old", "Sarah is 23 years old"))
}
