% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{stopifnot2}
\alias{stopifnot2}
\title{Wrapper around stopifnot allowing for a custom error message}
\usage{
stopifnot2(..., msg = NULL)
}
\arguments{
\item{...}{zero or more expressions to be evaluated}

\item{msg}{a custom error message to be printed if the expressions are not
all true}
}
\description{
Wrapper around stopifnot allowing for a custom error message
}
